#152 move ResampledTile to geotrellis-common
