+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/geotiff/package.scala
@@ -24,7 +24,7 @@ import org.apache.spark.util.AccumulatorV2
 import org.openeo.geotrellis
 import org.openeo.geotrellis.creo.CreoS3Utils
 import org.openeo.geotrellis.stac.STACItem
-import org.openeo.geotrellis.tile.TileGrid
+import org.openeo.geotrellis.tile_grid.TileGrid
 import org.slf4j.LoggerFactory
 import software.amazon.awssdk.core.sync.RequestBody
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
